requirements.txt
Babel==1.3
CherryPy==3.7.0
Faker==0.0.4
Flask==0.10.1
Flask-Babel==0.9
Flask-Classy==0.6.10
Flask-Login==0.2.11
Flask-Script==2.0.5
Flask-WTF==0.11
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.2
Werkzeug==0.10.4
aiocron==0.2
fake-factory==0.5.1
flask-mongoengine==0.7.1
futures==2.2.0
ipython==3.0.0
itsdangerous==0.24
mongoengine==0.9.0
numpy==1.9.2
pymongo==2.8
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
speaklater==1.3
wsgiref==0.1.2
